back-end/requirements.txt
# Core
Flask>=2.2,<3.0
python-dotenv>=1.0.0,<2.0.0

# Security and Performance
Flask-CORS>=4.0.0,<5.0.0
Flask-Limiter>=3.5.0,<4.0.0

# DB - Multi-database support
mysql-connector-python>=8.1.0,<9.0.0
psycopg2-binary>=2.9.0,<3.0.0
# SQLite is built into Python, no additional package needed

# Firebase Admin SDK
firebase-admin>=6.2.0,<7.0.0

# Google AI / Gemini
google-generativeai>=0.3.0,<1.0.0

# Firestore (if used directly)
google-cloud-firestore>=2.16.0,<3.0.0

# Auth
google-auth>=2.29.0,<3.0.0

# HTTP client (used by Firebase / GCP libs)
requests>=2.28.0,<3.0.0

# Gunicorn server for production
gunicorn>=20.1.0,<21.0.0
